requirements.txt
faiss-cpu
langchain
beautifulsoup4
PyPDF2
openai
flask
flask-socketio